refactored chooseOptimalSize out of Camera2BasicHandling and into a new CameraUtils class
